Replace entry id from auto increment to UUID (#384)
* initial attempt for uuid fix

* test problems after changing uuid

* The test should not assume the folder iteration ordering

* Fix super type finding in ontology generator.

Added the types that are contains subscript such as BaseGroup[Entry]. Make sure BaseGroup is in the super type list.

* Fix the return order of links and groups.

Use a sub-iterator instead of the id set joining methods. This may also speed up some entry lookup.

Links and groups are now returned in the order of creation. And entries are no longer sorted by the tid, now only sorted by the object types.

* Fix the return orders in multi pack.

* Fix ontonotes test.

* Fix openie test and allennlp tests.

* fix annotation removing

* Type expanding error

* mypy issues

* fix writer test

* Fix the remove in allennlp processor.

* Weird typo in setup.py

* docstring fix

* Fixing docstrings and some function names.

* mypy

* error in refactor